+++ JGL_MeshLoader/source/render/render_base.h
@@ -34,7 +34,7 @@ namespace nrender
     FrameBuffer() : mFBO { 0 }, mDepthId {0}
     {}
 
-    virtual void create_buffers(float width, float height) = 0;
+    virtual void create_buffers(int32_t width, int32_t height) = 0;
 
     virtual void delete_buffers() = 0;
 
@@ -48,8 +48,8 @@ namespace nrender
     uint32_t mFBO = 0;
     uint32_t mTexId = 0;
     uint32_t mDepthId = 0;
-    float mWidth = 0;
-    float mHeight = 0;
+    int32_t mWidth = 0;
+    int32_t mHeight = 0;
   };
 
   class RenderContext
